Renamed DataProcess class files to pep8 convention
