+++ ApertureMapModelTools/DataProcessing/__init__.py
@@ -7,9 +7,9 @@ Last Modifed: 2016/02/26
 #
 """
 #
-from .__Percentiles__ import Percentiles
-from .__EvalChannels__ import EvalChannels
-from .__Histogram__ import Histogram
-from .__HistogramRange__ import HistogramRange
-from .__HistogramLogscale__ import HistogramLogscale
-from .__Profile__ import Profile
+from .percentiles import Percentiles
+from .eval_channels import EvalChannels
+from .histogram import Histogram
+from .histogram_range import HistogramRange
+from .histogram_logscale import HistogramLogscale
+from .profile import Profile
