+++ ApertureMapModelTools/DataProcessing/base_processor.py
