+++ ApertureMapModelTools/DataProcessing/eval_channels.py
@@ -8,7 +8,7 @@ Last Modifed: 2016/10/25
 """
 #
 from ..__core__ import _get_logger
-from .__BaseProcessor__ import BaseProcessor
+from .base_processor import BaseProcessor
 logger = _get_logger(__name__)
