+++ ApertureMapModelTools/DataProcessing/histogram.py
@@ -9,7 +9,7 @@ Last Modifed: 2016/10/25
 """
 import scipy as sp
 from ..__core__ import _get_logger, calc_percentile
-from .__BaseProcessor__ import BaseProcessor
+from .base_processor import BaseProcessor
 logger = _get_logger(__name__)
